requirements.txt
pip>=25.0.1
audioread==3.0.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
dnspython==2.7.0
Flask==3.0.3
fonttools==4.54.1
greenlet==3.1.1
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.40.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.7.1
msgpack==1.1.0
numba==0.57.1
numpy==1.23.5
obspy==1.4.1
packaging==24.1
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
pymongo==4.10.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.3.0
scipy==1.11.4
setuptools==75.1.0
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.5.0.post1
SQLAlchemy==1.4.54
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
whitenoise==6.7.0
flask
flask-socketio
geopy
apscheduler
